Change default factoring behavior
Needed to move around *lots* of stuff in order for this to work. You can
still get an auto-factored expression if you use one of the <Op>F
constructors.

Now the Buf, Not, Xor, and Implies operators can be stored as native
expressions. The following will still auto-simplify:

* Buf(Number)
* Not(Number)
* Not(Literal)

There is now a hierarchy of expressions:
* Canonized: All terms have equal length = |support|
* Flattened: Nested OR/AND operators eliminated
* Factored: Consists of literals, and sums/products of literals
* Simplified: Numbers eliminated, and excess terms absorbed