Update to version 0.2.0
Now testing with Python3

Changed the OrAnd arguments (xs) to a list instead of a set.
As a result, I had to re-work a *lot* of code. The motivation
here is that I want to prioritize the speed of substitutions.

Figuring out idempotence and doing term absorption turned out
to be quite a bit harder, but substitution is now *much* faster.