Fix naive_sat_one, and implement naive_sat_all
The current implementation of naive_sat_all will iterate over all
satisfying assignment points, but whenever it reaches a TRUE leaf, it
does not check whether all variables in the support are part of the
point.

For example, if you traverse the equation f = a * b + a * -b,
naive_sat_all will yield a single solution at {a: 1} instead of yielding
both {a: 1, b:0}, {a:1, b:1}.

This actually has some nice properties, so I think I'll keep it this
way.