Update truth table string format
Previously, a truth table looked like this:

inputs: x2 x1 x0
000 0
001 0
010 0
011 1
100 0
101 1
110 1
111 1

Now, that same truth table will look like this:

x2 x1 x0
 0  0  0 : 0
 0  0  1 : 0
 0  1  0 : 0
 0  1  1 : 1
 1  0  0 : 0
 1  0  1 : 1
 1  1  0 : 1
 1  1  1 : 1

The main difference is that I got rid of the "inputs: ..." string.
Now, all the columns are aligned with the input assignments in the same
column as their corresponding input.
Since there is more whitespace now, I added the semicolon as a visual
aid for the output column.