Bugfix + new unit-tests for cardinality_constraints
ExactlyN(), which calls AtLeastN()+AtMostN was using the same auxvar-name
for newly introduced variables which seemed to be source of trouble,
discovered by the new unit-tests.

The new unit-tests imitate the ones of NHot & co. BUT are only valid
in a "is_feasible"-setting like explained in the sources.

See: #128